data: add dump function for documents
* Computes `_text` field for provisionActivity and series to documents, to display unstructured statement from the structured data.
