+++ rero_ils/modules/documents/api.py
@@ -25,6 +25,7 @@ from invenio_circulation.search.api import search_by_pid
 from invenio_search.api import RecordsSearch
 
 from .models import DocumentIdentifier
+from .utils import publication_statement_text, series_format_text
 from ..api import IlsRecord
 from ..fetchers import id_fetcher
 from ..minters import id_minter
@@ -128,3 +129,15 @@ class Document(IlsRecord):
         if self.harvested:
             cannot_delete['others'] = dict(harvested=True)
         return cannot_delete
+
+    def dumps(self, **kwargs):
+        """Return pure Python dictionary with record metadata."""
+        dump = super(Document, self).dumps(**kwargs)
+        provision_activities = dump.get('provisionActivity')
+        for provision_activity in provision_activities:
+            provision_activity["_text"] = \
+                publication_statement_text(provision_activity)
+        series = dump.get('series')
+        for series_element in series:
+            series_element["_text"] = series_format_text(series_element)
+        return dump
